requirements.txt
pypdf
cryptography
langchain
chromadb
pytest
boto3
langchain_community
langchain_ollama
langchain_chroma
streamlit
python-dotenv
openai
langchain-openai